Split apart gpstop from gpstart.
